feat: support vuepress v2
BREAKING CHANGE: remove vuepress-plugin-code-copy
"You cannot use other plugins in your plugin anymore, which avoids lots of
potential issues caused by plugin nesting. If your plugin depends on other plugins, you should list
them in the docs."

fix #41
