+++ .gitignore
@@ -5,4 +5,7 @@ coverage
 node_modules
 docs/.vuepress/dist/
 
+.cache
+.temp
+
 yarn.lock
