+++ docs/.vuepress/include-plugin.js
@@ -1,0 +1,13 @@
+
+module.exports = () => {
+    return {
+        name: 'vuepress-plugin-include',
+        extendsMarkdown: (md) => {
+            md.use(require('markdown-it-include'), {
+                root: './docs/',
+                includeRe: /<\[include\](.+)/i,
+                bracesAreOptional: true,
+            })
+        },
+    }
+}
