+++ docs/zh/README.md
@@ -1,1 +1,1 @@
-<[include](../README.md)
+<[include](../README-zh_CN.md)
