+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "vuepress-plugin-demo-code",
-    "version": "0.6.0",
+    "version": "1.0.0-0",
     "description": "📝 Demo and code plugin for vuepress",
     "main": "src/index.js",
     "files": [
@@ -33,8 +33,7 @@
     "dependencies": {
         "codesandbox-import-utils": "^2.2.3",
         "markdown-it-container": "^3.0.0",
-        "prismjs": "^1.23.0",
-        "vuepress-plugin-code-copy": "^1.0.6"
+        "prismjs": "^1.23.0"
     },
     "devDependencies": {
         "@babel/core": "^7.14.6",
@@ -42,6 +41,7 @@
         "@commitlint/cli": "^12.1.4",
         "@commitlint/config-conventional": "^12.1.4",
         "@vue/test-utils": "^1.2.1",
+        "@vuepress/plugin-search": "^2.0.0-beta.27",
         "all-contributors-cli": "^6.20.0",
         "babel-core": "^7.0.0-bridge.0",
         "babel-eslint": "^10.1.0",
@@ -66,7 +66,7 @@
         "rimraf": "^3.0.2",
         "standard-version": "^9.3.0",
         "vue-jest": "^3.0.7",
-        "vuepress": "^1.8.2",
+        "vuepress": "^2.0.0-beta.27",
         "vuepress-plugin-smooth-scroll": "^0.0.10"
     },
     "keywords": [
