+++ src/enhanceAppFile.js
@@ -1,5 +1,6 @@
+import { defineClientAppEnhance } from '@vuepress/client'
 import DemoAndCode from './DemoAndCode.vue'
 
-export default ({ Vue }) => {
-    Vue.component('DemoAndCode', DemoAndCode)
-}
+export default defineClientAppEnhance(({ app }) => {
+    app.component('DemoAndCode', DemoAndCode)
+})
